# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountExchange } from './accountExchange';

export interface Account {
  id: number;
  name: string;
  exchange: AccountExchange;
  ownerName: string;
  bankName: string;
  /** @nullable */
  bankUrl?: string | null;
  apiKeySet?: boolean;
  isActive: boolean;
  /** @nullable */
  balance?: number | null;
  /** @nullable */
  frozenBalance?: number | null;
  /** @nullable */
  totalVolume?: number | null;
  /** @nullable */
  totalProfit?: number | null;
  /** @nullable */
  completedTrades?: number | null;
  createdAt: string;
}