# lib/api-zod/src/generated/types/accountInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { AccountInputExchange } from './accountInputExchange';

export interface AccountInput {
  name: string;
  exchange: AccountInputExchange;
  ownerName: string;
  bankName: string;
  bankUrl?: string;
  apiKey?: string;
  apiSecret?: string;
}