# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * P2P Trading Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSide } from './orderSide';

export interface Order {
  id: number;
  accountId: number;
  /** @nullable */
  accountName?: string | null;
  /** @nullable */
  exchange?: string | null;
  side: OrderSide;
  asset: string;
  fiatCurrency: string;
  price: number;
  minAmount: number;
  maxAmount: number;
  /** @nullable */
  availableAmount?: number | null;
  /** @nullable */
  paymentMethod?: string | null;
  isActive: boolean;
  /** @nullable */
  completedTrades?: number | null;
  /** @nullable */
  completionRate?: number | null;
  /** @nullable */
  exchangeOrderId?: string | null;
  createdAt: string;
}